ffc/requirements.txt
django-imagekit
django-extensions